Function : Data Table 구조 변경 (3)
- 변경된 몬스터 Data Table 데이터들이 제대로 불러와지는 것 구현 & 테스트 완료.

+++ Assets/3_Scripts/Data/TableData/MonsterTable/MonsterTableClasses.cs
@@ -77,7 +77,7 @@ namespace MonsterTableClasses
         public int ID;
         public int attribute;
         public float multiplier;
-        public float effect;
+        public int effect;
         public float maintainTime;
         public float coolTime;
         public float defaultDamage;
@@ -90,7 +90,7 @@ namespace MonsterTableClasses
         public int ID;
         public int attribute;
         public float multiplier;
-        public float effect;
+        public int effect;
         public float maintainTime;
         public float coolTime;
         public float defaultConditionValue;
